Points 2 and 4 of TODO list
2) Now wallet client creates a Block only having a Timestamp and append
to it transactions without processing them; in networking (future) he
can then serialize the Block and broadcast it to other peers.

4) Added transaction fees: 1% of the value that the receiver would get
is sent to the miner, so a field 'miner' has been added to the
Transaction class and Chain.addBlock, Transaction.processTransaction,
Transaction.genesisTransaction and Block.mineBlock take the miner as
input. So, the sender must ensure that the value sent minus the
transactionFee associated to it is equal to the value that the receiver
expects. This could be accomplished in the Wallet.sendFunds method but
it would result in strong coupling: if transaction fee changes, I should
change both Wallet and Transaction classes. The solution to that issue
was to add a static field to Chain class called 'transactionFee': now
the transactionFee is charged to the sender in the Wallet.sendFunds
method.