Point 1 of TODO
Now it is possible for a Wallet to send funds to more receivers in a
single block, as long as the sender has a good pool of UTXOs; to solve
the problem of the pool of UTXOs, to send any list of values to any list
of receivers (where the sum of the values doesn't exceed the total
balance), in theory a wallet could make self-transactions to split, for
example, a big UTXO in small UTXOs, but it would suffer from transaction
fees. So I didn't implement this logic: the application throws an
exception if it's not possible to match the list. The user has to
manually split his UTXO if he really needs to make the transactions all
in a single block.